config: split CoS validators out of schema_validators.go
Move the CoS scheduler rate / buffer-size / forwarding-class leaf
validators (ValidateRate, ValidateByteSize, ValidateCoSTransmitRateTail,
ValidateCoSShapingRateTail, validateCoSRateTail, validateCoSPercentValue,
coSRateSiblingSuppliesValue, ValidateByteSizeOrPercent,
parsePercentWithSuffixStrict, validateForwardingClassRef) into a new
schema_validators_cos.go sibling.

Refs #4421 (codex-173 #1): schema_validators.go is a cross-domain
validator grab-bag (1159 LOC, 47 top-level decls). Pure code motion --
declarations moved verbatim (doc comments + bodies unchanged); only the
per-file package/import headers are synthesized. The generic value
primitives (LeafValidator, ValidateEnum, ValidateInteger[Min],
ValidatePercent, duration/wire const bounds) stay in schema_validators.go.
